Wrap all math operations in a check
This should resolve #9.
