+++ src/lang/math.rs
@@ -1,12 +1,13 @@
-use super::datatype::BoxInt;
+use super::interpreter::BoxInt;
 
 pub fn modulo<T: BoxInt>(a: T, b: T) -> Result<T, String> {
     if b.is_zero() {
-        return Err("Cannot use 0 as a modulus".to_string());
+        return Err("Modulo caused invalid value".to_string());
     }
 
-    if a * b < T::zero() {
-        Ok(b + a % b)
+    if a.checked_mul(&b).ok_or("Modulo caused invalid value")? < T::zero() {
+        Ok(b.checked_add(&(a % b))
+            .ok_or("Modulo caused invalid value")?)
     } else {
         Ok(a % b)
     }
@@ -16,7 +17,11 @@ pub fn inv_modulo<T: BoxInt>(a: T, b: T) -> Result<T, String> {
     let x = modulo(a, b)?;
     let mut n = T::one();
     while n < b {
-        let mod_result = modulo(n * x, b);
+        let mod_result = modulo(
+            n.checked_mul(&x)
+                .ok_or("Inverse modulo caused invalid value")?,
+            b,
+        );
         if mod_result.is_ok() && mod_result.unwrap().is_one() {
             return Ok(n);
         }
@@ -25,12 +30,4 @@ pub fn inv_modulo<T: BoxInt>(a: T, b: T) -> Result<T, String> {
     }
 
     return Err(format!("{} is not invertible", a));
-}
-
-pub fn divide<T: BoxInt>(a: T, b: T) -> Result<T, String> {
-    if b.is_zero() {
-        Err("Cannot use 0 as a divisor".to_string())
-    } else {
-        Ok(a / b)
-    }
 }
