+++ src/lang/mod.rs
@@ -1,4 +1,3 @@
-mod datatype;
 mod expression;
 pub mod interpreter;
 mod math;
